rx fm

+++ include/hacktvlib.h
@@ -14,7 +14,7 @@
 class HackTvLib{
 public:
     using LogCallback = std::function<void(const std::string&)>;
-    using DataCallback = std::function<void(const int16_t*, size_t)>;
+    using DataCallback = std::function<void(const int8_t*, size_t)>;
 
     HackTvLib();
     ~HackTvLib();
